// ng2-components/ng2-alfresco-search/src/services/alfresco-search.service.d.ts
/*!
 * @license
 * Copyright 2016 Alfresco Software, Ltd.
 *
 * Licensed under the Apache License, Version 2.0 (the "License");
 * you may not use this file except in compliance with the License.
 * You may obtain a copy of the License at
 *
 *     http://www.apache.org/licenses/LICENSE-2.0
 *
 * Unless required by applicable law or agreed to in writing, software
 * distributed under the License is distributed on an "AS IS" BASIS,
 * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
 * See the License for the specific language governing permissions and
 * limitations under the License.
 */
import { Observable } from 'rxjs/Rx';
import { AlfrescoAuthenticationService, AlfrescoApiService } from 'ng2-alfresco-core';
export declare class AlfrescoSearchService {
    authService: AlfrescoAuthenticationService;
    private apiService;
    constructor(authService: AlfrescoAuthenticationService, apiService: AlfrescoApiService);
    getNodeQueryResults(term: string, options?: SearchOptions): Observable<any>;
    private getQueryNodesPromise(term, opts);
    private handleError(error);
}
export interface SearchOptions {
    skipCount?: number;
    maxItems?: number;
    rootNodeId?: string;
    nodeType?: string;
    include?: string[];
    orderBy?: string;
    fields?: string[];
}